+++ src/beacon.rs
@@ -2,7 +2,7 @@
 // Copyright (C) 2015-2021  Dennis Schwerdel
 // This software is licensed under GPL-3 or newer (see LICENSE.md)
 
-use ring::digest;
+use crate::crypto_ring::digest;
 
 #[cfg(unix)]
 use std::os::unix::fs::PermissionsExt;